+++ test/taskmanager/http/handler/AbstractHandlerTest.java
@@ -2,13 +2,18 @@ package taskmanager.http.handler;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.params.provider.Arguments;
 import taskmanager.manager.InMemoryTaskManager;
 import taskmanager.http.HttpTaskServer;
+import taskmanager.utiltask.Epic;
+import taskmanager.utiltask.Subtask;
+import taskmanager.utiltask.Task;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.URI;
 import java.net.URL;
+import java.util.stream.Stream;
 
 public abstract class AbstractHandlerTest {
     protected HttpTaskServer httpTaskServer;
@@ -33,4 +38,26 @@ public abstract class AbstractHandlerTest {
         connection.setRequestMethod(method);
         return connection;
     }
+
+    static Stream<Arguments> provideTaskData() {
+        return Stream.of(
+                Arguments.of(new Task("Test Task 1", "Test Description 1")),
+                Arguments.of(new Task("Test Task 2", "Test Description 2"))
+        );
+    }
+
+    static Stream<Arguments> provideEpicData() {
+        return Stream.of(
+                Arguments.of(new Epic("Test Epic 1", "Epic Description 1")),
+                Arguments.of(new Epic("Test Epic 2", "Epic Description 2"))
+        );
+    }
+
+    static Stream<Arguments> provideSubtaskData() {
+        return Stream.of(
+                Arguments.of(new Subtask("Test Subtask 1", "Subtask Description 1", 1)),
+                Arguments.of(new Subtask("Test Subtask 2", "Subtask Description 2", 1))
+        );
+    }
+
 }
