+++ test/taskmanager/http/handler/HistoryHandlerTest.java
@@ -1,13 +1,11 @@
 package taskmanager.http.handler;
 
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import taskmanager.utiltask.Task;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
-import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -23,10 +21,4 @@ public class HistoryHandlerTest extends AbstractHandlerTest {
         assertEquals(200, connection.getResponseCode());
     }
 
-    static Stream<Arguments> provideTaskData() {
-        return Stream.of(
-                Arguments.of(new Task("Test Task 1", "Test Description 1")),
-                Arguments.of(new Task("Test Task 2", "Test Description 2"))
-        );
-    }
 }
