+++ test/taskmanager/http/handler/PrioritizedHandlerTest.java
@@ -1,14 +1,12 @@
 package taskmanager.http.handler;
 
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import taskmanager.utiltask.Task;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.time.LocalDateTime;
-import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -24,10 +22,4 @@ public class PrioritizedHandlerTest extends AbstractHandlerTest {
         assertEquals(200, connection.getResponseCode());
     }
 
-    static Stream<Arguments> provideTaskData() {
-        return Stream.of(
-                Arguments.of(new Task("Test Task 1", "Test Description 1")),
-                Arguments.of(new Task("Test Task 2", "Test Description 2"))
-        );
-    }
 }
