+++ test/taskmanager/http/handler/TaskHandlerTest.java
@@ -1,7 +1,6 @@
 package taskmanager.http.handler;
 
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import taskmanager.utiltask.Task;
 
@@ -10,19 +9,11 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class TaskHandlerTest extends AbstractHandlerTest {
-
-    static Stream<Arguments> provideTaskData() {
-        return Stream.of(
-                Arguments.of(new Task("Test Task 1", "Test Description 1")),
-                Arguments.of(new Task("Test Task 2", "Test Description 2"))
-        );
-    }
 
     @ParameterizedTest
     @MethodSource("provideTaskData")
